requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.1.0
inflection==0.5.1
pytz==2021.1
sqlparse==0.4.1
